backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
feedparser==6.0.11
newspaper3k==0.2.8
lxml_html_clean==0.1.1
transformers==4.37.2
torch==2.2.0
python-dateutil==2.8.2
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1